Adding a debug menu to monitor different game values

+++ SFML-Game-Test/Player.h
@@ -16,6 +16,6 @@ private:
 	std::list<Bullet*>* bulletList;
 	sf::CircleShape circle;
 	sf::RectangleShape line;
-	sf::Vector2f mapPosition;
+	sf::Vector2i mapPosition;
 	sf::Clock clock;
 };
